Starting to clean up javascript
