+++ includes/functions.php
@@ -971,8 +971,8 @@ function meta_noindex()
 function meta_favicon()
 {
 	$favicon_location = BASE_URI . 'img/favicon/';
-	echo '<link rel="icon" href="' . $favicon_location . 'favicon-32.png" />';
-	echo '<link rel="apple-touch-icon-precomposed" href="' . $favicon_location . 'favicon-152.png">';
+	echo '<link rel="icon" href="' . $favicon_location . 'favicon-32.png" />' . "\n";
+	echo '<link rel="apple-touch-icon-precomposed" href="' . $favicon_location . 'favicon-152.png">' . "\n";
 }
